requirements.txt
# Core dependencies
aiogram>=3.13.0,<4.0
aiosqlite>=0.19.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
APScheduler>=3.10.0,<4.0.0
aiofiles>=23.0.0

# Optional: для будущего использования PostgreSQL
# asyncpg>=0.29.0

# Optional: для распределенного кэша
# redis>=5.0.0
# hiredis>=2.2.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.5.0
